requirements.txt
python-box
pytest
requests
ipython
